Code/backend/requirements.txt
# Cancer Risk Stratification Model - Dependencies

# OCR Libraries
pytesseract>=0.3.10
Pillow>=10.0.0

# NLP and Transformers
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.13.0

# ML Libraries
xgboost>=2.0.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Note: Tesseract OCR engine must be installed separately:
# - Ubuntu/Debian: sudo apt-get install tesseract-ocr
# - macOS: brew install tesseract
# - Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
